feat :wrench: Add docker-comopose.yml
Docker Compose is a tool that facilitates the management of applications consisting of multiple services, containers, and configurations. It allows you to define and run multi-container applications using a YAML configuration file instead of having to use individual Docker commands for each container. This simplifies the creation, configuration, and deployment of complex applications that involve multiple interdependent components.