feat :sparkles: logic operators and=Joins query clauses with a logical AND retrns all documents that match the condition of both clauses.
not= Invert the effect of query expression and retirns documents that do not match the query expression.
nor= Joins query clauses with a logical NOR retirn all documents that fail match both clauses.
or= JOins query clauses with a logical OR returns all documents that match the conditions of either clauses.